requirements.txt
Flask
requests
python-dotenv
feedparser
Telethon
